What I added:
I set up the basic GUI structure we will use for now and made some buttons that work.
I also made a really crappy trending panel to show an example of how to set up the other panels.
Try to use that as you build your own crappy panels just to see if this thing works and we are sending all the info needed thru the right channels.

Just to help...basic tutorial

- Create a class extending JPanel that will show the info
- Set the properties like i did with TrendingPanel
- set the parameters for the constructor to accept the needed info to display
- in the ViewMain class create a method that creates the JPanel type class you made and passes along the needed info.
- have that new JPanel type object set to the borderlayout.CENTER
- In the controller class create a method that uses the twitter object to collect the data and send it to the ViewMain method you made
- In the controller class in the switch statement at the bottom call the method you just created in the case most appropriate to what you are doing.
- check it to see if it works by running it and clicking the button.
- Don't forget in the ViewMain method you made to end it will " updateUI(); "

if you have any problems follow the path of the trending line like i did.